Change CategoryDto Structure
move categorydto into one folder
