+++ Models/Category/CategoryUpdateDto.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
-namespace init_api.Models
+namespace init_api.Models.Category
 {
-    public class SecCategoryUpdateDto
+    public class CategoryUpdateDto
     {
         [Required]
         public string? Name { get; set; }
