+++ Models/Category/SecCategoryAddDto.cs
@@ -1,13 +1,14 @@
 using System.ComponentModel.DataAnnotations;
-namespace init_api.Models
+namespace init_api.Models.Category
 {
-    public class ThirdCategoryAddDto
+    public class SecCategoryAddDto
     {
         [Required]
-        public Guid UUID { get; set; }
         public string? Name { get; set; }
         public string? Icon { get; set; }
+        public Guid UUID { get; set; }
         [Required]
         public Guid ParentUUID { get; set; }
+        public ICollection<ThirdCategoryAddDto> Children { get; set; } = new List<ThirdCategoryAddDto>();
     }
 }
