+++ Models/Category/SecCategoryUpdateDto.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
-namespace init_api.Models
+namespace init_api.Models.Category
 {
-    public class CategoryUpdateDto
+    public class SecCategoryUpdateDto
     {
         [Required]
         public string? Name { get; set; }
