+++ Models/Category/ThirdCategoryAddDto.cs
@@ -1,14 +1,13 @@
 using System.ComponentModel.DataAnnotations;
-namespace init_api.Models
+namespace init_api.Models.Category
 {
-    public class SecCategoryAddDto
+    public class ThirdCategoryAddDto
     {
         [Required]
+        public Guid UUID { get; set; }
         public string? Name { get; set; }
         public string? Icon { get; set; }
-        public Guid UUID { get; set; }
         [Required]
         public Guid ParentUUID { get; set; }
-        public ICollection<ThirdCategoryAddDto> Children { get; set; } = new List<ThirdCategoryAddDto>();
     }
 }
